modified the UC2program to UC3

+++ AddressBookProblems/ContactDetails.cs
@@ -13,7 +13,7 @@ namespace AddreessBook
         public string address { get; set; }
         public string city { get; set; }
         public string state { get; set; }
-        public int zipcode { get; set; }
+        public long zipcode { get; set; }
         public long PhoneNumber { get; set; }
         public string email { get; set; }
     }
